admin_ng: increase resources for calico on wikikube clusters
In T343900 the ML team investigated an issue with kubelet high
latencies for ExecSync operations. It turned out it was due to
the readiness probe, running `/usr/bin/check-status -r`, and
getting timeouts due to a constant heavy cpu throttling.

Increase the resources dedicated to kube-controller and typha
pods, since they are the most throttled ones.

Bug: T343900
Change-Id: I8d29749be8cfd9aa10a750c3a2605ba9d4a659eb