changeprop: allow retries for liftwing streams with 502 responses
Sometimes the LiftWing's istio gateway pods return 502s to Changeprop
(very low percentage of the overall traffic) for unknown (yet)
reasons. It would be nice to avoid loosing the events and retry,
since the 502 probably indicates an issue with the TCP connection
between the Envoy on the Istio gateway to the envoy on the model
server.

Change-Id: I5f6e1a9782ba60affb218b42e739b69f159a4db6