Remove duplicate definition of oidc client secret from datahub
We found out that the AUTH_OIDC_CLIENT_SECRET variable was being
created twice in the datahub-frontend chart. It was defined once
specifically by name and again when it steps through each of the
available secrets in values.config.private.

This commit removes the specific declaration and keeps the other.

Bug: T305874
Change-Id: I2c38bfd3677ab9c86cbf82e2a1d9dec60e5643be