Move jaeger from admin_ng to aux services
This moves the aux-k8s-eqiad jaeger deployment from being an admin_ng
component to being a usual service deployable without requiring root
etc.

As running things like this is the purpose of the aux cluster I think it
makes more sense to threat it as actual workload rather than a
kubernetes system component.

Bug: T344253
Change-Id: Ieefccf9ece15136afaf8cc479e31baebbd050488