datahub: set preferred oidc jwt algotithm
We have so far verified that authentication on the IDP side is okay and that we do
receive a signed id token. The challenge lies in how datahub is validating the token received.
We must use the same ID token JWS algorithm used by the IDP to validate the token received.

Bug: T305874
Change-Id: I79f6e39fd0b50c35380592a09b6777c0c13f1e9b