datahub: Add the oidc scope
We have so far verified that authentication on the IDP side is okay and that we do
receive a signed id token. We need to now add the profile scope to the AUTH_OIDC_SCOPE which
will allow us to get the prefered_username, the value which we use as the AUTH_OIDC_USER_NAME_CLAIM
that shall allow us to uniquely authenticate each user per WMF standards.

Bug: T305874
Change-Id: Ib04b173b879205614f6c0f8f887e5993f4288d70