ml-services: tune knative's container concurrency settings
In T344058#9086843 we discovered that "containerConcurrency" is
not really used if a lower soft limit is provided (via annotations).
Istead of using a hard limit (that causes requests to be buffered
by the Knative Activator pods IIUC) we just use the soft limit
to hint the autoscaler to add more instances if needed.

Raise the soft limits too since by default Knative scales up
earlier than the limit, at the moment we configure 85% of it
(to smooth the scale up transition timings and avoid overloads
due to a raise in traffic).

For the moment a hard limit seems dangerous, we may re-introduce
it in the future if needed.

Switch the autoscaling target metric to "rps" (supported by
KPA) to have a more friendly metric to compare with Istio rps
metrics in our dashboards.

Bug: T344058
Change-Id: Ic7161e58e98c65d7083e2476407b21ac2574e66e