datahub: add oidc production settings
We have so far verified that authentication on the IDP side is okay on staging.
Next we need to have the settings on production, this change disables JAAS based authentication and
enables the oidc settings already configured, to be used in production and adds idp servers to
the egress rules. This also removes the hardcoded values used in testing, and separates
production and staging settings.

Bug: T305874
Change-Id: I729689e9275562288bbf82ed27d3738b84668384