admin_ng: allow webhook to watch istio/knative ns and increase resources
The webhook changes applied in the last change to ease big knative
deployments didn't work, so for safety it is better to allow the webhook
to check the istio and knative namespaces.

Increase the webhook's resources to avoid throttling during upgrades,
likely causing delays.

Increase also the controller's resources to avoid throttling.

Change-Id: Ib26ec2ac753d0049e375efed87050e30c84f9d16