Update the maximum message size in kafka for eventstreams
We have increased the maximum message size in kafka-jumbo to 10 MB in
order to permit the mediawiki content enrichment flink job to pass
larger messages.

Since evenstreams is a consumer of this kafka topic it will need to be
configured with the larger message size, otherwise it will fail to
consume whenever it comes across a message larger than its currently
configured maximum.

Bug: T344688
Change-Id: I18484022d09a24a1d35e398ce9e70c03c8e065f8