charts: add the python-webapp chart
The ML team created a while ago the fastapi-app chart to support
ores-legacy, but while creating another chart for recommendation-api-ng
(using uwsgi but hopefully fast-api in the future) we realized
that the new chart would be almost identical, if not for bits like
command/image/etc.. that can be easily customizable under helmfile.d.

The idea is to introduce python-webapp, move ores-legacy to it and
finally deprecate fastapi-app.

This change is basically a copy of the fastapi-app chart with a minimal
modification of Chart.yaml's values.

Change-Id: I5f5c5cedbc57515ea234acaa62a0cf451f7d5343