update_version: drop python 3.5, 3.6. Add 3.9
ruamel.yaml has dropped support for python 3.5.

Python 3.5 and 3.6 are end of life and not included in any Debian
distribution anymore.

Keep python 3.7 for now since it is still in Debian Buser.
Add python 3.9 which is in Debian Bullseye.

Python 3.11 from Bookworm can't be added yet since the CI image does not
have it yet (T342019).

Change-Id: I0c45bbffee3e87da8046ea6374c0232939dd6755