services: increase concurrency in ORESFetchScoreJob's changeprop cfg
After the switch from ORES to LiftWing in the MW ORES extension,
we started seeing a backlog of changeprop jobs causing issues like
delayed scores etc.. Increase the current concurrency to 30 to
see if the current 20 value is the culprit.

Lift Wing's processing time is stated in T346175#9169094, the issue
is mostly related to enwiki and wikidata, and probably due to
the fact that before the switch we heavily cached and primed ORES Redis cache
(for the moment we don't have the same thing for Lift Wing).

Last but not the least, Lift Wing is based on microservices (one model
for each service), meanwhile with ORES we could do batch score calls.
This means that we make more requests to Lift Wing now, that in turn
mean more jobs to process.

Bug: T346175
Change-Id: Idd95b423a0685d7305ce8d677fc01537644275b0