ml-services: remove old eswikiquote and eswikibooks models
Remove old deployments of eswikibooks and eswikiquotes as they are no longer in use.
After the change in the hostnames for some wikis the following deployments are being removed:
eswikiquotewiki and eswikibooks as they have been replaced by deployments that do not have the wiki suffix
in order to be consistent with wikiID from MediaWiki configuration.

Bug: T342266, T345850
Change-Id: Ic5ecc14460681fc7bdafd34e0b2e2affbd765102