mediawiki: add support for a prometheus-statsd-exporter
As we're trying to finally undeploy graphite, we want to migrate
MediaWiki to use prometheus-statsd-exporter instead.

To this end, we add a prometheus-statsd-exporter sidecar that can
be enabled on demand in k8s. Only peculiarity we need to keep in mind
is I chose a fixed port for the statsd-exporter, port 9102, so
configuration of MediaWiki should proceed accordingly.

Bug: T343025
Change-Id: I3d276cd91098865aa4f678ae81b7e2f292a25852