ml-services: fix recommendation-api-ng readiness probe failure
When other parameters besides `spec` are queried,
the API takes longer to return a result which causes
the readiness probe to time out. This results in the
API query hanging without returning a result.

To address this issue, this patch increases the
readiness probe `timeoutSeconds` to accommodate the
API delay and prevent the API query from terminating mid-way.

Bug: T347475
Change-Id: Ib7b60d3b6e578c19e6e569792062884a7df64a82