rest-gateway: strictly order route definitions
Using a map here is easier on the eyes, but we can't be guaranteed
that this will be rendered in the order that we've requested. Enforce
ordering by switching to a list.

This is of particular note for the wikifeeds entries where there are
specific cases that should be matched before the general cases.

Change-Id: I3f083169f85496929651d502bef4dc74d269852e