ml-services: increase recommendation-api-ng uwsgi workers
When other parameters besides 'spec' are queried,
the API takes longer to return a result which causes
the k8s readiness probe to time out. This results in the
API query hanging without returning a result.

We run an investigation in T347475#9218749 that showed
resource constraints could be affecting the performance
of the endpoint.

To address this issue, this patch increases the
uwsgi workers and cpu to improve the application's
performance.

Bug: T347475
Change-Id: If7952303bc2763ef52e39556fd06faef3d82bdd7