rdf-streaming-updater: dse: Rename release to wikidata
We used the terms `wikidata` and `wdqs` somewhat interchangably
between the paths and the release name. Rather than having both
names standardize on `wikidata`. This was chosen because it's
easier to see the difference between wikidata and commons vs
wdqs and wcqs.

Some references to wdqs still remain. Particularly for the exported
topics I left it as wdqs primarily due to context. In the context
of the helm chart or stored resources it should be clear that wikidata
refers to the query service, but for topics discoverable through kafka
it's less clear that wikidata refers specifically to the query
service.

Change-Id: I33dc94b73cc53d8ee07647a4121ce2cfa0465121