api-gateway: add Content-type in the CORS' allowed headers settings
The JS apps calling Lift Wing may need to set the Content-Type header
to signal that the data sent via POST is JSON.

The error that the community is seeing:
https://github.com/WikiEducationFoundation/WikiEduDashboard/pull/5501#issuecomment-1753576067

This change adds the support for Content-Type to the common
CORS config. It shouldn't affect any other client, but just
allow the necessary headers.

Bug: T348511
Change-Id: I73a4d3c2c12695a7597ace89d37e989f9c69dc30