cirrus updater: topic-prefix-filter must be plain string prefix
This was initially going to be a regular expression to be applied to
the topics, but then we realized it could be much simpler a a string
prefix. Remove the regex's and provide only simple prefix strings.

Additionally updates the url's used for schema download. In testing
these need the jsonschema suffix as well so that appropriate
urls are constructed.

Bug: T347075
Change-Id: Ieb94ce5baf9cfdd89276b826340b56a85b1d9c18