cirrus updater: Configure kafka communication to use SSL
The updater is timing out while trying to communicate with kafka.
We've verified the ports are open all the way through, suspect that
the problem is that kafka doesn't have any ability to auto-detect
the connection type.

Resolve by telling it to use ssl explicitly, to match what our
kafka servers provide on port 9093.

Change-Id: Ifbd1ea68586f3c4377570b5aaea7436fbe7bfed4