flink-app: Configure internal metrics port
When running a flink app the admin ui reports the metrics are missing,
and pod logs regularly complain about being unable to establish a
flink-metrics connection. Much like the other ports, default to a
constant value and configure the appropriate networkpolicy.

Not 100% sure that we can get away with a single port instead of a
range here, but we can start here and adjust as necessary. To that
end this is initially optional and only configured in the cirrus
service. It can migrate to a default once verified working.

Change-Id: Id2a5c2e318ed3bab9bed1cf4029c76096c995c40