cirrus updater: Ensure all output topics exist
The consumer is failing to start because the event stream config says
both eqiad and codfw topics for the update topic exist, but we only
have the eqiad on in staging. Using the topic prefix filter didn't
help, the metadata lookup must be happening prior to filtering
by the event utilities library.

To resolve temporarily point the producer at the codfw prefixed topic
so that it will be auto-created.

Bug: T347075
Change-Id: Ic402a3e0fd0598fb80ecf0650c0157610db924bd