cirrus updater: Turn on two med-large wikis in the producer
Now that the updater is able to start and process basic testwiki
events, start running more events through the system. This only
turns on the producer side, the consumer will not fetch or attempt
to write these out to elasticsearch. A little more prep work is
necessary to send the writes to relforge, and this will start
feeding the partial upate events into the system.

Change-Id: I07d31db0195e317846cf5d44dd7ae6507488d060