cirrus updater: Re-enable the .* route for mwapi
While getting this running we ran into an issue with the http routes
and elasticsearch and hacked around it by disabling the .* rule. After
running a few tests in the application it looks like this approach
should work, and the problem before had to do with trailing slashes
not being a part of the HttpHost definition.

Change-Id: I862e188ed938c65995bfa9b3ee594225b668f13a