changeprop: refactor templating for Kafka producer/consumer settings
Instead of hardcoding kafka settings in the chart, split the kafka
producer/consumer settings into two parts:
- shared config
- dedicated config
The shared config is meant to be shared between changeprop instances,
meanwhile the dedicated config is something that can be tuned
with instance-specific values. The idea is to have a default set
of values shipped with the chart, and allow to customize the rest
when needed.

While doing it:
- linger.ms set to environment specific instances, rather than
  being a shared value. The idea is that every instance will need
  to set its own value, based on the workload and perf requested.
- The 'ssl' flag disappears and the TLS options are by default
  added to values.yaml (it is not a default everywhere). We have
  specific Beta values so it should be ok even for deployment-prep.
- Add a new "enabled" flag to the Jobqueue's config, instead of
  relying on the fact that the config is populated.
- Add a new fixture for the Jobqueue use case.

Bug: T348950
Change-Id: I0255990a421f959db5956d6acceb18da359df541