changeprop: set num_workers to zero
The current pstree layout for changeprop staging is:
node─┬─node───573*[{node}]
     └─134*[{node}]

Service runner creates an extra worker process that sits there
doing nothing, with a lot of threads that can be easily removed.

With num_workers=0:
node───573*[{node}]

It probably made sense on bare metal to have a nanny/worker,
but I don't see it being useful in the container world.

Bug: T348950
Change-Id: I18ec9bd9820872ba67ec9689cbd7c89cd475b117