rest-gateway: change how AQS URLs enforce wikimedia.org domain
Previously we checked the Host header to ensure that we only answered
for aqs services on wikimedia.org. We do this to avoid polluting the
cache with identical data for all of our sites. However, this is a
digression from how we've done things historically, only using the
/wikimedia.org/ path in our URLs. Some of our internal services (like
mediawiki and wikifeeds) don't set the header and in some cases
setting the header would be slightly arduous. In order to compromise,
match *only* on the domain wikimedia.org in the URL path, and drop the
requirement for the Host header to be set

Bug: T348731
Bug: T349517
Change-Id: Ie5a02119214978b0dee23dcf36643e96c13d1fb0