mediawiki: clarify comment on swift egress rules
For thumbnail pregeneration, we currently reach out to the local
swift in the current datacenter directly. We do have egress enabled
to those clusters because we do activate an envoy listener for them,
even if we still didn't migrate mediawiki-config to use it.

Make the comment declaring so more explicit because it currently tricked
me into thinking I was mistaken back when I wrote it.

Change-Id: Ib002b1cde0faae07c32d169f53245688bd5fa45d