Copy version from production service
NOTE: it seems that this service was running this version in production
but this was not mirrored in the source code.  When I went to deploy I
saw it in the diff and I'm not exactly sure what to do.  I'll deploy
this way for now and see if it breaks anything (service is currently not
serving correct data anyway)

Change-Id: Ia29d18fb2044c0c5a090352b5bb15f2ed3d309ac