api-gateway: Add env variables required for envoy SDS
File based SDS in envoy requires --service-cluster and --service-zone to
be set in envoy - which we do in the entrypoint.sh if the environment
variables are set.

I used the same defaults we use in modules/mesh/deployment.yaml here.
Also dropped SERVICE_IDENTIFIER, as that is no longer used.

Bug: T300033
Change-Id: Ide22776fe9b90b667d038d1291784fe08ed8388b