mcrouter: add vanila chart
This is an intermediate patch for the mcrouter so
to keep track of any changes made to run it as a DaemonSet.

The chart includes templates that are not needed, however
due to some dependencies, sextant installs them. We can
clean up the chart after we have sorted them out.

The actual chart can be found
in Ic852c39ac2b6dc7adb477182aa63e226a3e81633

Bug: T346690
Change-Id: I22c49d3c94c644f050cbc1724a3dbe641635814e