cirrus updater: Expand test deployment to prod+cloudelastic
The writes to the cloudelastic cluster are moving to the new
updater pipeline, eventually all writes will move but this
is the first production test deployment. Cloudelastic has been
chosen as the first target as it has the lowest risk if the
updates are incomplete.

To ensure no edits are lost this needs to be deployed and verified
working prior to the mediawiki-config patch that disables the old
update pipeline.

Bug: T352335
Change-Id: If85ce9618d7ee0d3e01d1409a9c2566cb0d619fb