mesh: Ship new configuration templates
Why:
Using wmf-ca-certificates for the service-mesh disallows talking to
services that are using Let's Encrypt issued certificates.
ca-certificates as a file is a superset including public CAs as well as
our private ones

What:
This is the no-op copy file change. The actual change will come in the
subsequent request

Bug: T352906
Change-Id: If7e788168f9cfbb3e42d93ddd9dbcb04b79c7050