recommendation-api: update monitoring config
The new statsd config will be needed to match the new service runner
metrics (2.8+) that we are going to deploy as part of
the nodejs 18 upgrade.
Add also a new prometheus flag to allow the usage of native prometheus
metrics. The port used by the Prometheus exporter will be the same one
used for the statsd one.

Bug: T205870
Change-Id: I403011bfb12253c4de77ef62e3d1e2598ad37182