ml-services: bump CPUs to compare with Research team benchmarks
While working on optimizing response performance for the
article-descriptions model-server the Research team shared
benchmarks that were run on 8 CPUs. This patch bumps up the
CPUs for this model-server in the experimental namespace to
match the Research team's benchmark environment so that we
can compare our results with their findings.

Bug: T353127
Change-Id: I613824eb84bc07dde26fd3f5f7f48a88fdbf2da8