admin_ng: force coredns to resolve to A records in ml-staging
We are currently testing a more 'transparent' proxy setup for
the istio mesh on ml-staging. The idea is to be able to write
Python code that can open HTTP connections to 'en.wikipedia.org'
without explicitly mention that we want to proxy to
api-ro.discovery.wmnet.

Why:
This would reduce the extra scaffolding/config needed when moving
from local testing (MLOps) to staging/production.

What:
Due to issues like https://github.com/istio/istio/issues/46625
the following happens:
- The Python code in a KServe pod makes a HTTP call to
  en.wikipedia.org
- Since we don't have the istio dns-proxy deployed, the domain
  is resolved by coredns that may return a AAAA record.
- With an A Record, the subsequent TCP connection started from
  the Python code is intercepted by the iptables chains and
  it gets correctly re-routed to the local envoy proxy,
  that knows what to do (namely proxy to api-ro.discovery.wmnet).
- With an AAAA record, the TCP connection gets stuck into a SYN
  state since it apparently bypasses the iptables rules.
The ml-staging cluster has ipv6dualstack enabled.
The trick to bypass the Istio CNI issue is to force Core DNS to
skip AAAA records, forcing them to A.

Bug: T353622
Change-Id: I48e3c1dee7a145a1e7328d8cf8ecdfdcb2066a10