mw-api-int: replicas x1.3
Add 27 replicas in preparation for the planned mobileapps traffic
increase. This should be about right to accomodate all mobileapps
traffic, but just in case, the traffic increase will happen in two
steps (from the current 75% first to 90% and later 100%).

The additional CPU requests would be a bit tight in eqiad with the
current number of nodes, so I will add more nodes before merging.

Bug: T350846
Change-Id: I2bd69f5f8742c0383f9d75662575a1c10859d427