mediawiki: Support one-off jobs
This adds support for a second way to run MediaWiki, as part of a
one-off maintenance job (the sort previously run on the maintenance
server with `mwscript`).

When .Values.mwscript.enabled is off, this change is a no-op.

When .Values.mwscript.enabled is *on*, we don't launch a normal MW
Deployment, and instead launch a Job which launches MWScript.php plus
the args provided from the values file. The deployment is partly
stripped down, consisting only of that php container plus mcrouter,
envoy, rsyslog, and the statsd exporter.

When the php container terminates, the sidecar controller cleans up the
whole pod (as we've set the "pod.kubernetes.io/sidecars" annotation on
it). The pod survives for one hour after that (ttlSecondsAfterFinished)
for convenient log-checking.

Bug: T341553
Change-Id: I558cd56d6787516925ba32f866fd349bc0fe6866