Add helmfile for running MediaWiki one-off jobs.
The NAME_TOKEN environment variable is different for each run, which
prevents naming collisions.

This normally isn't applied directly; instead it's used through a Python
wrapper script, which passes an appropriate NAME_TOKEN (including the
username and script, for convenience) and correctly-formulated values
for any given script execution. That script is in a separate patch, to
be installed on the deployment hosts.

Bug: T341553
Change-Id: Ia519393acbeab563f6a8e7d7603d7ee759682c3f