base.meta: Remove dependency on the mesh module
Why:
Having a dependency from the base module to the mesh module is prone to
cause various problems (and actually does!).

The biggest one is that it can cause dependency chain incompatibilities
as the base module, a module that is prudent to only source from other
modules due to it's ubiquity and nature, ends up sourcing the mesh
module with specific version pinning, meaning the user won't be able to
decide which version of the mesh module they can use in their chart, but
rather get pinned on a specific version (at best), or met with difficult
to diagnose sextant errors in case of a version mismatch

The other one is a cyclic dependency between 2 diffirent modules:

* base.meta sources mesh.name
* mesh.name sources both mesh.configuration and mesh.certificate which
  source mesh.name

Now all of above are essentially hardly coupled and need to change in
tandem, seldomly a good pattern.

What:

This patch introduces a backwards incompatible change, removing the
usage of mesh.name.annotations from the base module. This can't be put
somewhere else, it should be included by the charts directly, which
means chart owners that want to use the mesh need to take action

Due to this backwards incompatibility, we follow Semver and bump
base.meta to 2.0.0

Add the removed part into the scaffolding

Change-Id: I2b0b0c24a9c78d5ec9ff6123ab76673d25985512