cxserver: Remove all kademlia support from chart
Why:
We never managed to get this working in production. This is because in
the case of an environment where ephemeral instances of the software
exist by default, as is the case in kubernetes, it's very easy to end up
in a pathological case where the Kademlia DHT is badly constructed with
multiple split brains where pods are grouped into small DHTs structures
that don't talk to each other. This could have been mitigated if the
limitation node_module that powers this was properly maintained but it
is not.

What:
Remove all kademlia functionality from the chart. It was never working
in the kubernetes world anyway

Bug: T355167
Change-Id: I716baad9b4d60d8b8c17d875f321104a2dec75a6