Update the spark-operator image name and version
We have recently changed the way that we build spark, in order to
support several different minor versions at once.

This commit updates the chart to use the new naming scheme, along with
the latest version of that chart so that it picks up the lates Java 8
JRE.

Although we have three versions of the spark-operator available, we are
sticking with version 3.3 in this commit.

Bug: T354273
Change-Id: I052c14c263ba25536119b2941ba977db9bb67b3d