cirrus: Expand production deployment wikis
The production producers now process the full set of wikis, no direct
limitations are applied. This has been running the same config in
staging for some time.

The production consumer-cloudelastic has had its wiki list updated to
include around 25% of the events we expect to ingest for the full
pipeline. This has been running at 100% in staging for more than a
month (now disabled), and at 1% in prod for a couple days.

Bug: T352335
Change-Id: Ic8a051c8d6ed965b9358e1e111d6e4a0f38b06b7