mw-debug: Enable tracing with 100% sampling
Why:
We want to be able to trace 100% of the traffic that reaches mw-debug.

What:
Add tracing.enabled and tracing.sampling stanzas. This should trace all
of the traffic going through the service proxy, whether it's terminated
at the TLS terminating local_service or egressing via the service mesh
listeners

Bug: T351566
Change-Id: I0a105874f0ccc68ae93880c617ded39a604a3171