linkrecommendation: Bump memory limit by 200Mi
Why:
Since the introduction of an alerting system in T266216,
linkrecommendation-internal has pods that are pretty close to the limit.

What:
Bump the limit by 200Mi. This number isn't created via some extensive
research. It's ~30% increase and we got 2/8 pods past the alerting
threshold. Naively speaking, this should be sufficient

Bug: T266216
Change-Id: If35793413ab4907785cd04a49319a26196a4bfee