cirrus updater: Introduce backfill releases
As part of reindexing operations we need the ability to reprocess
recent update streams and re-apply them to specific wikis. The helm
side of this is provided by defining a paired backfill release for
each normal consumer release that has the same configuration but
applies bounds to the kafka inputs. The reindexing process will need
to submit the backfill and wait for it to finish as part of the
process.

Bug: T356303
Change-Id: Iaa3e31128c54267cb59337f47964e71bb2d66a19