cirrus: Add conditional -backfill releases (take two)
The first stab at this didn't quite work as we want, if a backfill
is deployed and the regular deployment is run it looks like helmfile
wants to undeploy the -backfill release. The hope here is that if
we only define the releases that are relevant to the context then
helmfile won't attempt to do anything to -backfills during normal
releases, and vice-versa.

Additionally slightly refactored how the backfill release is
templated. The `backfill` template is now a complete definition
and not just a thing we copy out of. This helps keep the actual
release definitions very simple and direct.

Change-Id: I3f061060da8afa738eb9604f601980f55df2562e