Add an nginx reverse proxy to superset to help with serving static assets
We wish to serve the static assets for superset from an nginx reverse
proxy and it is OK to serve these files without authentication.

This change adds the nginx pod running on port 8080 and configures this
as the next hop after the TLS termination. Nginx receives the plaintext
request and tries the URL. If it is present on the file system it is
returned, but if it is not, it is forwarded to the app running on port
8090.

We configure nginx to write both the error_log and access_log to stdout
in order to log the traffic.

Bug: T357890
Change-Id: I7e31071745e4140d4145a3c611bddcefb8046d1b