Improve the superset nginx reverse proxy configuration
We needed to set some additional nginx directives in order to be able
to use it as a reverse proxy.

In addition, we have updated the container versions because we rebuilt
the superset containers. We now include the assets with the backend
container, because it requires them in order to include static content
in dynamic responses.

Bug: T357890
Change-Id: Id35787fe95f2482fe3ff0a94a060e58c08d84999