superset: assign Superset roles from LDAP groups
We define a mapping of LDAP group (cn) to Superset roles.
As we set auth_roles_sync_at_login=True, this means that a Superset
user roles will get updates at each refresh.

We also make sure to update the Superset User model with the details
provided to us by the OIDC provider. For example, if an existing user
had an email of brouberol@email.notfound and the OIDC provider returns
`{..., 'email': 'brouberol@wikimedia.org'}, we make sure to automatically
update the Superset User record with these details.

Bug: T297120
Bug: T353794
Change-Id: I6d0c6b2ea3e4709caf691382ff4e9bdd1ece1de1