cirrus: Enable consumer-cloudelastic writes to all wikis
Removes the wikis limit for the cloudelastic consumer. With this the
consumer will be performing 100% of production writes against the
cloudelastic cluster. A related patch will disable writes from the
cirrus side after this is deployed.

Worth remembering that the input wikis do not include private wikis,
as private wikis do not generate our input events. Cloudelastic
contains only public wikis, so this really is 100%. For that same
reason no private wiki filter is being provided here.

Bug: T358518
Change-Id: I662174f783ad233a8c5b36066aa7b013e15fe4bc