shellbox: add PHP + Apache timeout settings
PHP-FPM:
- max_execution_time
- process_control_timeout

Apache:
- TimeOut directive

These currently use the defaults set in the image (if set at all).
However, it makes sense to me that for shellbox they all should be
about equal to the request timeout.

None of the existing deployments change php.timeout, so this should not
change behaviour (but see below). It will be relevant for videoscalers.

Note that this change is *not* a complete no-op, even though it
shouldn't change behaviour in practice: the existing defaults
mostly match the default php.timeout, except for max_execution_time,
which is 120s currently and this change will set it to 60s. This should
be irrelevant, as FPM__request_terminate_timeout kills everything
after 60s anyway, but it is a change in the settings.

Bug: T357309
Change-Id: I74e518595f75a747bc60f129353a713ae4cb431c