superset: allow usage of DATA_CACHE_CONFIG
In previous superset versions, DATA_CACHE_CONFIG was deemed a security issue,
because a user wihout read rights on the original data would we allowed to see
the cached data.

We're making no-op changes to the chart that will enable us to re-evaluate that
behavior along with the CACHE_IMPERSONATION and CACHE_QUERY_BY_USER feature flags
on superset next.

Bug: T358753
Change-Id: I7011d16fb6ecb8f9d46d1b75a70d40d4e38a56c6