mw-api-int: Hold eqiad back on resources
The eqiad wikikube cluster is lacking a bit in resources compared to
codfw. As such, the mw-api-int capacity raise needs to be held back
until more appservers are migrated. This is not a problem for now as
this increase is for mobileapps traffic, which is almost entirely in
codfw.

However, we need to consolidate the two environments as soon as
resources allow in case a failover is needed.

Bug: T356497
Change-Id: I498c40f050f0d6f8e47b6297e6a2f2d3d52ed737