Increase tls timeout for mw-jobrunners to fix upload jobs
Upload jobs (AssembleUploadChunks & PublishStashedFile) need to
calculate SHA1 hashes of large files, and do other longish operations
on files. Uploads of large files on commons have been broken due to
the jobs timing out after three minutes. This appears to be due to
the TLS timeout having a default for MW instances of 3 minutes.
Prior to switching to k8s, the job runners had a tls timeout a
few seconds greater than the apache timeout for the job runner
(a108d093eec) so add the same here. After this change, the lowest
timeout for job running would be the change-propagation timeout
which defaults to 7 minutes.

Bug: T358308
Change-Id: I82156553e2d81b7af99c6c21efa4fcf9e73de321