eventstreams: change default num_workers to 0.
Service runner creates an extra worker process that sits there
doing nothing, with a lot of threads that can be easily removed.

It probably made sense on bare metal to have a nanny/worker,
but I don't see it being useful in the container world.

This patch aligns the configuration to changeprop and eventgate.

Bug: T359051
Change-Id: I8363f7200314a3dd05a613637b878d746f6c48f7