cirrus: Check backfill status prior to reindexing
In the case where a backfill is already deployed and a reindex was
requested we will do the full reindex, and then fail the backfill.
It seems much more helpful to fail-fast instead of waiting until we've
done that work and missing the backfilling.

While here also update for an error seen in prod, which prompted the
change to error handling, where querying the job status from k8s gets
a KeyError but only immediately after deployment. Seems like a race
with the initialization routines, simply expect to get a missing key
there.

Change-Id: I745329c339379a3e7ad5b84dd0afb0a779ad77fd