changeprop: Move staging to mw-api-int
Why?

As part of the mw-on-k8s migration, internal clients of the
api_appserver bare metal cluster must be moved to use mw-api-int.

What?

As a first step to validate correct functionality, move the staging
release of changeprop to use mw-api-int as a mediawiki API backend.
Add egress rules to mw-api-int.

Bug: T360767
Change-Id: Icafb534c99130a3323a4416655f3bde86fc5433a