Fix pod label selector for external-services network policy
The pod label selector was not matching any deployed pod, as it didn't respect
our standard label conventions.

These are the labels we have on a pod deployed by the superset chart, with release=staging

```
  labels:
    app: superset
    pod-template-hash: 5d89cc49cf
    release: staging
    routed_via: staging
```

Nowhere does `name` appear. This is something we'd setup during our test phase that
should not have made it to production.

Bug: T331894
Change-Id: If47ca33f35a446b0ff5cd1b87d499980019e99ae