cirrus: More reliable reporting of reindexing status
We've seen a case where the wikidata backfill reports via the flink
apis that it is finished, but the flinkdeployment state still thinks
everything is running. It seems that there are some edge cases which
make flinkdeployment unreliable. Augment the flinkdeployment state
check with a query against the flink rest api to get the final status
and have a more reliable shutdown.

Additionally allows the script to end when the backfill reaches the
NOT_DEPLOYED state. This occurs when an operator manually destroys
the backfill release and will be considered a failure.

And, adds a wait for the application to become undeployed after
destroying it. Some wikis skipped reindexing when doing a full
cluster reindex as they saw the previous backfill still existing.

Change-Id: I293a08743a7c7a6ecfa6c57df0b97eef6afedeac