Migrate datahub to use external-services for CAS IDP
We wish to avoid having to use hard-coded IP addresses for services such
as the CAS-SSO IDP servers within our helmfile deployments. To this end,
we have now developed a module calles base/external-services which
allows us to reference the services by name.

In the case of Datahub, we want to migrate the CAS-IDP settings first
because the CAS service has recently been switched to new servers,
causing Datahub to break.

This commit intends to fix that by using the new network policies for
CAS IDP.

Bug: T331894
Change-Id: Ie69d88f656d627868273094e58ead5b6a944388a