charts/kserve-inference: Wire up generated network policy for LW services
This adds a new version of modules/base/external-services-networkpolicy

Since the kserve-inference services on LiftWing have all set `app` to
`kserve-inference` and instead use `app-wmf` to distinguish individual
services, the new version allows the user to set
`external_services_appname` to the desired value. If this is unset, the
old hardcoded `app` is used.

Bug: T360428,T359423
Change-Id: I23ae8c6fd283e210f38cafaf09fe3c7ca16750e1