Create a new aqs-http-gateway chart
At the moment, we have two separate charts for the two types of AQS
service, named druid-http-gateway and cassandra-http-gateway.

We have a requirement in the Commons Impact Metrics endpoint, which is
currently in development for AQS, to be able to communicate with both
types of backend data store.

This commit adds a new chart called aqs-http-gateway, which is a merged
version of the two existing charts.

We expect to be able to migrate all of the existing AQS endpoint
services to use this new aqs-http-gateway chart in due course, then
remove the two discrete versions.

Bug: T360531
Change-Id: I706dfe457b24c61700facc7c58c6ab4878f7c1eb